#93 refactor: API 수정 및 리뷰 반영
